+++ NEWS.md
@@ -1,6 +1,8 @@
 # GPTCM 0.0.1
 
 * Add modeling of the proportions through the common parametrization of Dirichlet regression, i.e. without reference category
+* Add options of IGamma and Gamma prior for Weibull's shape parameter kappa
+* Change the prior variance of zeta's intercept from fixed value 10 to an IGamma prior
 
 # GPTCM 0.0.1